refactor: extract link state machine into useLinkWorkflow hook
